LinearRegression/pipeline/requirements.txt
azure-ai-ml==1.10.0
azure-identity==1.13.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.3
scipy==1.14.1
joblib==1.4.2
python-dotenv==1.0.0
